+++ sbin/tinyshell.lua
@@ -1,5 +1,5 @@
 local _VERSION = "test release"
-local gpu = system.display.getGPU()
+local gpu = os.simpleDisplay.gpu
 local filesystem = system.kernel.getModule("filesystem")
 local keyCodes = dofile("/lib/keyboard.lua").keys
 local screenWidth, screenHeight = gpu.getResolution()